Deleted a bunch of stuff.
There was a lot in this repo which wasn't my thing and which had
gone unmaintained the instant it arrived. I will never allow that
sort of thing to happen again. At this time there are at most two
files of interest in this repository, ignoring readmes and licenses.
Now those are the only two.