Dropped -MaxPermSize=384m and -Dfile.encoding=UTF8.
On java8 the first generates a warning every time the jvm
is started and I'm pretty sure the latter is unnecessary
now that apple is no longer foisting macro-macro-man on us.

I tried to only drop those options on java8 and leave java7
and previous alone, but some interaction I don't understand
meant the tests would not pass. Contributions welcome on
that point. You can always enumerate the options you want
so it's an inconvenience at worst.