Add a test for -java-home.
The script, of course, runs java at the end, so java_cmd must be an
executable for the test to run successfully, which in the tests is
managed by stubs.

Also, when a test is run it's from within a "myproject" directory.

So this test is dependent on those two things.

Related: paulp/sbt-extras#83