JAVA_HOME now also sets javaHome, quietly.
JAVA_HOME, in addition to setting the version of java to run the
launcher with, now also quiety sets javaHome.

In order to set it quietly the log level is changed, which broke the log
level commandline flags.  Seeing as JAVA_HOME is detected and javaHome
is set before any commandline flag is evaluated, by re-issuing any log
level flag as an sbt command the log levels set correctly again.