Drop OS X support for Travis..
This reverts the principle change in #129.

Continuing from where that commit message left off, the reason that the tests
never use Java is that's how bats works: nothing on the PATH is used.

Therefore there's no strong reason to test on both Linux and OS X as it's only
testing potential variations of whatever version of bash is running.

So dropping OS X support reduces the travis setup and recovers some speed in the
builds (hopefully back to ~40s from ~1m30).

Note OS X support might come back with #132.