Fix Cygwin support
Only Cygwin paths are used by every external tool used by the script with the exception of Java, which only accepts Windows paths.

The current state of the script evaluates `mkdir -p "${jar%/*}"` as the full Windows path, turning the sbt launcher jar into a directory. This causes the script to promptly exit with the error "Download failed. Obtain the jar manually and place it at ...".

This pull request proposes an additional variable, `$sbt_jar_jvm`, which stores the appropriate representation of `$sbt_jar` for use when invoking the JVM.

This addresses #198