Fix Cygwin support (alternative implementation)
Every external tool used by the script accepts only Cygwin paths, with the exception of the JVM, which accepts only Windows paths.

The current state of the script evaluates `mkdir -p "${jar%/*}"` as the full Windows path, turning the sbt launcher jar into a directory. This causes the script to promptly exit with the error `Download failed. Obtain the jar manually and place it at ...`.

This pull request proposes an additional function, `jar_file_jvm`, which converts `$sbt_jar` to the appropriate representation for use when invoking the JVM. The `jar_file` function will no longer perform this conversion.

This addresses #198